Carpeta de la práctica 2

+++ Práctica_1/ejercicio3.cpp
@@ -6,7 +6,7 @@ int main(void){
 	double c, f;
 	cout << "ingrese los grados centigrados a convertir" << endl;
 	cin >> c;
-	f = static_cast<double>(static_cast<double>(9)/static_cast<double>(5)) * c + 32.0;
+	f = static_cast<double>(9)/5 * c + 32.0;
 	cout << "el  equivalente en grados fahrenheit es:" << f << endl;
 	return 0;
 }
